Rename test files
Thils removes all the test_ prefix from test files, so they do not get
discovered by python unittest. This has several reasons:
* Tests should ony be started with the suite now (or manually which
  still works)
* I can't make the suite working with just running the suite file
  without the unittest module, so this is a hacky way of still allowing
  us to use `discover`
