+++ tests/PubchemInfo.py
