+++ tests/PubchemInfoNoInet.py
