+++ tests/annotateInchiRoutines.py
