+++ tests/parseMetaboliteInfoFromSBML.py
